api_gcs_old/pyproject.toml
[tool.poetry]
name = "infra-app-api"
version = "0.1.0"
description = ""
authors = ["Will Beebe"]
packages = [
  {include = "*", from="src"},
]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.0"
fastapi = "^0.111.0"
google-cloud-storage = "^2.17.0"
python-jose = "^3.3.0"
google-auth = "^2.31.0"
jsonschema = "^4.22.0"


[tool.poetry.dev-dependencies]
pytest = "^8.1.1"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
select = ["E", "F", "W", "I001"]
ignore = ["E501", "F541"]
fixable = ["W", "I001"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
